added length of children array to path
This makes the relative change much simpler both in code as well as in
architectur, because now the leave can skip nodes with correct
orientation BUT to less children and check the ancestors of this node
for matching properties and stay focused if there is none.

Former this was not possible since the leave did not know how many
children an ancetor of it had, so it had to probe it.
This resulted in multiple msg-cycles and if there was no fitting node
the former focus leave was again focused, bcause of the last focused
default.

Also:
- renamed variables to convey more meaning.
- removed error collection and display used for crude debbuging.