changed interface to include dedicated Methods
to circumvent the private batchMsg from bubbletea and thus to
preventing nested batchMsg's. This has the lovely side effect that the
the architecture is also becoming simpler, since multiple Msg structs are
unnecessary, like the following:

- InitIDs
- Start
- Ready

This is because each of there purpose was resolved through directly
returning multiple Cmd's and not having to squash them into one private
batchMsg.