changed methods to return errors directly instead of Cmd'S
since Cmd's are only intended for side-effect operations, and not for
(self-)communication. This bothers me a little bit, since i would prefer it if
there would be a way to listen for what is changing in the Model, still
i see that my solution was not beautiful or a good way to do it.
Now one has to check it by one self, through checking which methods one
calls.

Also made the Sort-interface methods public since its not anymore
necessary to make all methods, which change the Model, return a Cmd
which holds the information what changed.